refactor: Category And Channel 구조 변경
Closes 호환성 문제 발생. 해결중
