+++ src/main/java/com/sprint/mission/discodeit/Iterator/ChannelIterator.java
@@ -1,10 +1,6 @@
 package com.sprint.mission.discodeit.Iterator;
 
-import com.sprint.mission.discodeit.composit.Category;
-import com.sprint.mission.discodeit.composit.CategoryAndChannel;
-import com.sprint.mission.discodeit.composit.Channel;
-
-import java.util.ArrayList;
+import com.sprint.mission.discodeit.service.jcf.CategoryAndChannel;
 
 public class ChannelIterator implements Iterator {
     private CategoryAndChannel channel;
