+++ src/main/java/com/sprint/mission/discodeit/entity/Server.java
@@ -1,12 +1,6 @@
 package com.sprint.mission.discodeit.entity;
 
 
-import com.sprint.mission.discodeit.composit.Category;
-import com.sprint.mission.discodeit.composit.CategoryAndChannel;
-import com.sprint.mission.discodeit.composit.Channel;
-
-import java.util.LinkedList;
-
 public class Server extends BaseEntity{
     public Server(String id, String name) {
         super(id, name);
