--- src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -1,6 +1,0 @@
-package com.sprint.mission.discodeit.service;
-
-public interface ChannelService {
-
-
-}
